bit restructuring and blur effect on modal open
