+++ src/App.js
@@ -1,4 +1,4 @@
-import Home from "./pages/home";
+import Home from "./pages/HomePage";
 import { useEffect, useState } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import SplashScreen from "./components/SplashScreen";
@@ -53,7 +53,6 @@ function App() {
   return (
     <ThemeProvider theme={theme === "light" ? lightTheme : darkTheme}>
       <GlobalStyles />
-
       <div className="App">
         {loading ? (
           <SplashScreen />
